refactor(server): move out userWorkspacePermissions form twnety-shared as it's not the dto
