+++ packages/twenty-server/src/engine/metadata-modules/role/dtos/user-workspace-permissions.dto.ts
@@ -1,0 +1,6 @@
+import { UserWorkspace } from "src/engine/core-modules/user-workspace/user-workspace.entity";
+
+export type UserWorkspacePermissionsDto = Pick<
+  UserWorkspace,
+  'objectPermissions' | 'settingsPermissions' | 'objectRecordsPermissions'
+>;
