+++ packages/twenty-shared/src/types/index.ts
@@ -13,4 +13,3 @@ export { FieldMetadataType } from './FieldMetadataType';
 export type { IsExactly } from './IsExactly';
 export type { ObjectRecordsPermissions } from './ObjectRecordsPermissions';
 export type { ObjectRecordsPermissionsByRoleId } from './ObjectRecordsPermissionsByRoleId';
-export type { UserWorkspacePermissions } from './UserWorkspacePermissions';
